Implement TaskMask
Previously, assumptions were made in a lot of places, like TaskMap that
MAX_TASKS will always be 32. By implementing TaskMask, we can change it
to 8, 16 or 32 without anything breaking.
Currently its unsafe to go above 32, for example to 64 - as assumption
that whole map fits in usize on 32-bit system (u32) is still in place.

+++ whyos/src/scheduler.rs
@@ -7,7 +7,7 @@ use cortex_m_rt::exception;
 use critical_section::Mutex;
 use defmt::warn;
 
-pub type TaskMask = u32; // TODO: just an idea for now, should be used in bitmaps
+pub type TaskMask = u8; // FIXME: right now, it can't go above u32, because of active_tasks syscall
 pub const MAX_TASKS: usize = TaskMask::BITS as usize;
 pub const IDLE_TID: TaskId = unsafe { TaskId::new_unchecked(0) };
